backend/requirements.txt
Django==5.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.1.1
djangorestframework-simplejwt==5.5.1
cryptography==46.0.1
psycopg2-binary==2.9.10
redis==5.2.1
django-redis==5.4.0
whitenoise==6.9.0
gunicorn==23.0.0
dj-database-url==2.3.0
google-auth==2.40.3
requests==2.32.5